+++ odeon/commons/reports/report_binary.py
@@ -47,8 +47,9 @@ class Report_Binary(Report):
 
         dict_export['cms'] = cms_json
 
+        json_object = json.dumps(dict_export)
         with open(os.path.join(self.input_object.output_path, 'report_metrics.json'), "w") as output_file:
-            json.dump(dict_export, output_file, indent=4)
+            output_file.write(json_object)
 
     def to_md(self):
         """Create a report in the markdown format.